requirements.txt
tensorflow==2.15.0
numpy==1.26.4
pillow==10.2.0
scikit-learn==1.4.2
matplotlib==3.8.4
streamlit==1.37.1
